Add files via upload
Platform differs from other DAWs because it expands upon current workstation familiarities to make them more social and connected. Using Blockchain technology anyone with an account on-chain can claim an address, anything they make, instruments and songs they acquire, are all stored with that address and are accessible from anywhere via a connected device. Although other tools work in a similar fashion, none allow for universal access through link and id sharing, nor provide a way to give access to VSTs in a web3 based app.

The goal of this project is to decentralize the music industry and create user owned ecosystems in which music (in all forms - full tracks, sessions, stems, VSTs) can be shared, sold, bought, and played with an interface that is accessible from anywhere at any time.